fix(lint): API level and duplicate label errors in the rich text toolbar
- layout_marginHorizontal/Vertical 은 API 26 이고 minSdk 는 24 라, 구분선
  스타일은 네 방향을 따로 적습니다.
- 버튼과 그 색 견본이 스크린 리더에 똑같이 "Text colour" 로 읽혀서
  DuplicateCrowdInStrings 에 걸렸습니다. 견본 쪽을 "... swatch" 로 구분합니다.
  읽어주는 내용으로도 이쪽이 맞습니다.
- 0x66...... 은 Int 범위 안이라 toInt() 가 군더더기였습니다. 불투명한
  0xFF...... 쪽은 Long 이라 변환이 그대로 필요합니다.

+++ AnkiDroid/src/main/java/com/ichi2/anki/richtext/RichTextEditor.kt
@@ -223,11 +223,11 @@ class RichTextEditor(
          */
         val DEFAULT_HIGHLIGHTS =
             listOf(
-                0x66FFF59D.toInt(), // yellow
-                0x66A5D6A7.toInt(), // green
-                0x6690CAF9.toInt(), // blue
-                0x66F48FB1.toInt(), // pink
-                0x66CE93D8.toInt(), // purple
+                0x66FFF59D, // yellow
+                0x66A5D6A7, // green
+                0x6690CAF9, // blue
+                0x66F48FB1, // pink
+                0x66CE93D8, // purple
             )
 
         /**
